requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.7.1
pinecone-client==3.2.2
langchain==0.1.16
langchain-openai==0.1.3
# --- THIS IS THE FIX: The missing spellbook ---
langchain-pinecone==0.1.0
# --- END OF FIX ---